nominal value refactor
